# packages/openapi/src/gen/handlers/createAuthorHandler.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { CreateAuthorMutationResponse, CreateAuthor400 } from "../types/CreateAuthor.ts";
import { http } from "msw";

export function createAuthorHandlerResponse201(data: CreateAuthorMutationResponse) {
  return new Response(JSON.stringify(data), {
    status: 201,
      headers: {
      'Content-Type': 'application/json'
    },
  })
}

export function createAuthorHandlerResponse400(data?: CreateAuthor400) {
  return new Response(JSON.stringify(data), {
    status: 400,
  
  })
}

export function createAuthorHandler(data?: CreateAuthorMutationResponse | ((
        info: Parameters<Parameters<typeof http.post>[1]>[0],
      ) => Response | Promise<Response>)) {
  return http.post('/authors', function handler(info) {
    if(typeof data === 'function') return data(info)

    return new Response(JSON.stringify(data), {
      status: 201,
        headers: {
        'Content-Type': 'application/json'
      },
    })
  })
}