# packages/openapi/src/gen/handlers/createQuoteHandler.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { CreateQuoteMutationResponse, CreateQuote400 } from "../types/CreateQuote.ts";
import { http } from "msw";

export function createQuoteHandlerResponse201(data: CreateQuoteMutationResponse) {
  return new Response(JSON.stringify(data), {
    status: 201,
      headers: {
      'Content-Type': 'application/json'
    },
  })
}

export function createQuoteHandlerResponse400(data?: CreateQuote400) {
  return new Response(JSON.stringify(data), {
    status: 400,
  
  })
}

export function createQuoteHandler(data?: CreateQuoteMutationResponse | ((
        info: Parameters<Parameters<typeof http.post>[1]>[0],
      ) => Response | Promise<Response>)) {
  return http.post('/quotes', function handler(info) {
    if(typeof data === 'function') return data(info)

    return new Response(JSON.stringify(data), {
      status: 201,
        headers: {
        'Content-Type': 'application/json'
      },
    })
  })
}