# packages/openapi/src/gen/handlers/deleteAuthorHandler.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { DeleteAuthorMutationResponse, DeleteAuthor404 } from "../types/DeleteAuthor.ts";
import { http } from "msw";

export function deleteAuthorHandlerResponse200(data?: DeleteAuthorMutationResponse) {
  return new Response(JSON.stringify(data), {
    status: 200,
  
  })
}

export function deleteAuthorHandlerResponse404(data?: DeleteAuthor404) {
  return new Response(JSON.stringify(data), {
    status: 404,
  
  })
}

export function deleteAuthorHandler(data?: string | number | boolean | null | object | ((
        info: Parameters<Parameters<typeof http.delete>[1]>[0],
      ) => Response | Promise<Response>)) {
  return http.delete('/authors/:authorId', function handler(info) {
    if(typeof data === 'function') return data(info)

    return new Response(JSON.stringify(data), {
      status: 200,
    
    })
  })
}