# packages/openapi/src/gen/handlers/deleteQuoteHandler.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { DeleteQuoteMutationResponse, DeleteQuote404 } from "../types/DeleteQuote.ts";
import { http } from "msw";

export function deleteQuoteHandlerResponse200(data?: DeleteQuoteMutationResponse) {
  return new Response(JSON.stringify(data), {
    status: 200,
  
  })
}

export function deleteQuoteHandlerResponse404(data?: DeleteQuote404) {
  return new Response(JSON.stringify(data), {
    status: 404,
  
  })
}

export function deleteQuoteHandler(data?: string | number | boolean | null | object | ((
        info: Parameters<Parameters<typeof http.delete>[1]>[0],
      ) => Response | Promise<Response>)) {
  return http.delete('/quotes/:quoteId', function handler(info) {
    if(typeof data === 'function') return data(info)

    return new Response(JSON.stringify(data), {
      status: 200,
    
    })
  })
}