# packages/openapi/src/gen/handlers/getAuthorByIdHandler.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { GetAuthorByIdQueryResponse, GetAuthorById404 } from "../types/GetAuthorById.ts";
import { http } from "msw";

export function getAuthorByIdHandlerResponse200(data: GetAuthorByIdQueryResponse) {
  return new Response(JSON.stringify(data), {
    status: 200,
      headers: {
      'Content-Type': 'application/json'
    },
  })
}

export function getAuthorByIdHandlerResponse404(data?: GetAuthorById404) {
  return new Response(JSON.stringify(data), {
    status: 404,
  
  })
}

export function getAuthorByIdHandler(data?: GetAuthorByIdQueryResponse | ((
        info: Parameters<Parameters<typeof http.get>[1]>[0],
      ) => Response | Promise<Response>)) {
  return http.get('/authors/:authorId', function handler(info) {
    if(typeof data === 'function') return data(info)

    return new Response(JSON.stringify(data), {
      status: 200,
        headers: {
        'Content-Type': 'application/json'
      },
    })
  })
}