# packages/openapi/src/gen/handlers/getQuoteByIdHandler.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { GetQuoteByIdQueryResponse, GetQuoteById404 } from "../types/GetQuoteById.ts";
import { http } from "msw";

export function getQuoteByIdHandlerResponse200(data: GetQuoteByIdQueryResponse) {
  return new Response(JSON.stringify(data), {
    status: 200,
      headers: {
      'Content-Type': 'application/json'
    },
  })
}

export function getQuoteByIdHandlerResponse404(data?: GetQuoteById404) {
  return new Response(JSON.stringify(data), {
    status: 404,
  
  })
}

export function getQuoteByIdHandler(data?: GetQuoteByIdQueryResponse | ((
        info: Parameters<Parameters<typeof http.get>[1]>[0],
      ) => Response | Promise<Response>)) {
  return http.get('/quotes/:quoteId', function handler(info) {
    if(typeof data === 'function') return data(info)

    return new Response(JSON.stringify(data), {
      status: 200,
        headers: {
        'Content-Type': 'application/json'
      },
    })
  })
}