# packages/openapi/src/gen/handlers/updateAuthorHandler.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { UpdateAuthorMutationResponse, UpdateAuthor400, UpdateAuthor404 } from "../types/UpdateAuthor.ts";
import { http } from "msw";

export function updateAuthorHandlerResponse200(data?: UpdateAuthorMutationResponse) {
  return new Response(JSON.stringify(data), {
    status: 200,
  
  })
}

export function updateAuthorHandlerResponse400(data?: UpdateAuthor400) {
  return new Response(JSON.stringify(data), {
    status: 400,
  
  })
}

export function updateAuthorHandlerResponse404(data?: UpdateAuthor404) {
  return new Response(JSON.stringify(data), {
    status: 404,
  
  })
}

export function updateAuthorHandler(data?: string | number | boolean | null | object | ((
        info: Parameters<Parameters<typeof http.put>[1]>[0],
      ) => Response | Promise<Response>)) {
  return http.put('/authors/:authorId', function handler(info) {
    if(typeof data === 'function') return data(info)

    return new Response(JSON.stringify(data), {
      status: 200,
    
    })
  })
}