# packages/openapi/src/gen/handlers/updateQuoteHandler.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { UpdateQuoteMutationResponse, UpdateQuote400, UpdateQuote404 } from "../types/UpdateQuote.ts";
import { http } from "msw";

export function updateQuoteHandlerResponse200(data?: UpdateQuoteMutationResponse) {
  return new Response(JSON.stringify(data), {
    status: 200,
  
  })
}

export function updateQuoteHandlerResponse400(data?: UpdateQuote400) {
  return new Response(JSON.stringify(data), {
    status: 400,
  
  })
}

export function updateQuoteHandlerResponse404(data?: UpdateQuote404) {
  return new Response(JSON.stringify(data), {
    status: 404,
  
  })
}

export function updateQuoteHandler(data?: string | number | boolean | null | object | ((
        info: Parameters<Parameters<typeof http.put>[1]>[0],
      ) => Response | Promise<Response>)) {
  return http.put('/quotes/:quoteId', function handler(info) {
    if(typeof data === 'function') return data(info)

    return new Response(JSON.stringify(data), {
      status: 200,
    
    })
  })
}