# packages/openapi/src/gen/types/Author.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { Quote } from "./Quote.ts";

export type Author = {
    /**
     * @type integer | undefined, int64
    */
    id?: number;
    /**
     * @type string
    */
    firstName: string;
    /**
     * @type string
    */
    lastName: string;
    /**
     * @type string | undefined
    */
    bio?: string;
    /**
     * @type array | undefined
    */
    quotes?: Quote[];
};