# packages/openapi/src/gen/types/Quote.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { Author } from "./Author.ts";

export type Quote = {
    /**
     * @type integer | undefined, int64
    */
    id?: number;
    /**
     * @type string
    */
    text: string;
    /**
     * @type object
    */
    author: Author;
};